# procrastinator-backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Procrastinator API
  version: 1.0.0
  description: |
    The Procrastinator HTTP API. All operations live under the shared path
    prefix `/api/users/{userId}` (path tenancy): the user is resolved from the
    route and every resource is scoped to that user.

    Wire-shape conventions (Design D7):
    - Property names are `snake_case` (verbatim JSON keys).
    - Optional / omitempty fields are `nullable: true` and are NOT in `required`.
    - Money and price values are exact-decimal `string` (never a number).
    - `link_conflicting` is a required, non-nullable boolean (always present).
    - `lines` on an import batch is an `array` that may be `null`.
    - `metadata` on an asset is a non-nullable free-form object (never null).
    - Dates are `format: date`; timestamps are `format: date-time`.
    - Non-2xx responses use the `error` envelope.
servers:
  - url: /
    description: Base path (path tenancy is in the route prefix)
paths:
  /api/users/{userId}/documents:
    parameters:
      - name: userId
        in: path
        required: true
        description: The user the resource belongs to (path tenancy).
        schema:
          type: string
    get:
      operationId: listDocuments
      summary: List documents
      description: |
        Lists all documents for the user, joined with source metadata.
        Supports filtering by processing status and searching by filename.
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by processing status.
          schema:
            type: string
            enum: [processed, in_review, failed, asset_less]
        - name: q
          in: query
          required: false
          description: Case-insensitive substring match on source filename.
          schema:
            type: string
      responses:
        "200":
          description: The user's documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/document'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    post:
      operationId: uploadDocument
      summary: Upload a document
      description: |
        Uploads a document as `multipart/form-data`, runs the ingest pipeline,
        and returns the created asset. An optional `owner_household_id` scopes
        the document to a household (the requester must be a member); when
        absent the document is personal.
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                owner_household_id:
                  type: string
                  nullable: true
                note:
                  type: string
                  nullable: true
                  description: Optional free-text note (user directive) for the extraction.
              required:
                - file
      responses:
        "201":
          description: Created asset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/asset'
        "202":
          description: Held for review (confidence below threshold)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ingest_review'
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "413":
          description: Upload exceeds size limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "415":
          description: Unsupported file type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "422":
          description: No usable identity in document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "409":
          description: Duplicate document detected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/duplicate_report'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "502":
          description: Extraction failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /api/users/{userId}/documents/{id}:
    parameters:
      - name: userId
        in: path
        required: true
        description: The user the resource belongs to (path tenancy).
        schema:
          type: string
      - name: id
        in: path
        required: true
        description: The document id.
        schema:
          type: string
    delete:
      operationId: deleteDocument
      summary: Delete a document
      description: |
        Soft-deletes the document and detaches it from its asset (the asset
        remains, now asset-less-eligible). The linked asset is NOT deleted.
      responses:
        "204":
          description: Document deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /api/users/{userId}/documents/{id}/reprocess:
    parameters:
      - name: userId
        in: path
        required: true
        description: The user the resource belongs to (path tenancy).
        schema:
          type: string
      - name: id
        in: path
        required: true
        description: The document id.
        schema:
          type: string
    post:
      operationId: reprocessDocument
      summary: Reprocess a document
      description: |
        Re-runs the extraction pipeline for the document. An optional comment
        is persisted to data.user_directive and passed into the extraction
        prompt. Returns 409 when the document is already in-flight (in_review).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/reprocess_document_request'
      responses:
        "202":
          description: Reprocess accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/document'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "409":
          description: Document is already in-flight
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /api/users/{userId}/documents/{id}/keep:
    parameters:
      - name: userId
        in: path
        required: true
        description: The user the resource belongs to (path tenancy).
        schema:
          type: string
      - name: id
        in: path
        required: true
        description: The document id.
        schema:
          type: string
    post:
      operationId: keepDocument
      summary: Keep existing (resolve pending choice)
      description: |
        Resolves the document's pending_choice to keep_existing. No
        re-extraction runs. Returns 200 with the updated document.
      responses:
        "200":
          description: Kept existing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/document'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /api/users/{userId}/add:
    parameters:
      - name: userId
        in: path
        required: true
        description: The user the resource belongs to (path tenancy).
        schema:
          type: string
    post:
      operationId: addItems
      summary: Unified add (files and/or text)
      description: |
        Accepts one or more files (multipart) and/or a text body. Each item is
        auto-detected: statement files are routed to the ledger import pipeline
        (requiring `account_id`); all other items go through the document
        processing pipeline. Returns a uniform per-item outcome array.
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                text:
                  type: string
                  nullable: true
                account_id:
                  type: string
                  nullable: true
                  description: Required for statement items; scopes the import to a finance account.
      responses:
        "200":
          description: Per-item outcomes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/add_item_outcome'
        "400":
          description: Invalid request (e.g. statement without account_id)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "401":
          description: Missing or invalid user identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "413":
          description: Upload exceeds size limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "415":
          description: Unsupported file type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /api/users/{userId}/assets:
    parameters:
      - name: userId
        in: path
        required: true
        description: The user the resource belongs to (path tenancy).
        schema:
          type: string
    get:
      operationId: listAssets
      summary: List all assets
      description: Returns all assets visible to the user, ordered by created_at then id.
      parameters:
        - name: include_deleted
          in: query
          required: false
          description: Include soft-deleted assets in the result
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: The user's assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/asset'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /api/users/{userId}/assets/{assetId}:
    parameters:
      - name: userId
        in: path
        required: true
        description: The user the resource belongs to (path tenancy).
        schema:
          type: string
      - name: assetId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getAsset
      summary: Get a single asset
      description: Returns a single asset by id. Unknown or another user's assets fail with 404.
      parameters:
        - name: include_deleted
          in: query
          required: false
          description: Include soft-deleted assets in the result
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: The asset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/asset'
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    delete:
      operationId: deleteAsset
      summary: Soft-delete an asset
      description: |
        Soft-deletes an asset by setting its `deleted_at` timestamp. The asset is
        excluded from list, search, and get responses (404 on direct get) until
        restored or purged. Owner-scoped; unknown or another user's ids yield 404.
      responses:
        "204":
          description: No Content
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    patch:
      operationId: patchAsset
      summary: Patch an asset (user corrections)
      description: |
        Applies user corrections to an asset. At minimum supports `asset_category`
        (setting it marks the category as user-set and sticky against future
        inference). Owner-scoped; unknown or another user's ids yield 404.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/patch_asset_request'
      responses:
        "200":
          description: The updated asset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/asset'
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /api/users/{userId}/assets/{assetId}/restore:
    parameters:
      - name: userId
        in: path
        required: true
        description: The user the resource belongs to (path tenancy).
        schema:
          type: string
      - name: assetId
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: restoreAsset
      summary: Restore a soft-deleted asset
      description: |
        Restores a soft-deleted asset within the retention window, clearing
        `deleted_at`. The asset reappears in list/search with all prior fields
        and document links intact. Restoring beyond the retention window yields
        409. Owner-scoped; unknown or another user's ids yield 404.
      responses:
        "200":
          description: The restored asset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/asset'
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "409":
          description: Asset is beyond the retention window
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
  /api/users/{userId}/assets/{assetId}/documents:
    parameters:
      - name: userId
        in: path
        required: true
        description: The user the resource belongs to (path tenancy).
        schema:
          type: string
      - name: assetId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: listAssetDocuments
      summary: List documents for an asset
      description: |
        Returns all documents attached to an asset. The asset is looked up first
        so unknown or another user's assets fail with 404 instead of an empty list.
      responses:
        "200":
          description: The asset's documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/document'
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
components:
  schemas:
...
    error:
      type: object
      description: The standard error envelope for non-2xx responses.
      properties:
        error:
          type: string
      required:
        - error
...
    duplicate_report:
      type: object
      description: The structured 409 response when a duplicate document upload is detected.
      properties:
        code:
          type: string
          const: "duplicate"
        existing_document_id:
          type: string
          nullable: true
          description: The existing document id, or null when the duplicate source has no linked document (e.g. statement-classified sources).
        existing_source_filename:
          type: string
        existing_source_uploaded_at:
          type: string
          format: date-time
        existing_asset_id:
          type: string
          nullable: true
        prompt:
          type: object
          description: The reprocess/keep prompt fields for the UI modal.
          properties:
            reprocess_uri:
              type: string
            keep_uri:
              type: string
            expires_at:
              type: string
              format: date-time
            timeout_toast:
              type: string
          required: [reprocess_uri, keep_uri, expires_at, timeout_toast]
      required: [code, existing_source_filename, existing_source_uploaded_at, prompt]
...